feat: migrate from Blade templates to React for client-side rendering

+++ app/Http/Controllers/ProductController.php
@@ -10,22 +10,25 @@ use App\Models\Product;
 class ProductController extends Controller
 {
     //show product page
-    public function index() {
+    public function index()
+    {
         $products = Product::orderBy('created-at', 'DESC')->get();
-        return view('products.list', [
-            'products' =>  $products
-        ]);
-
-
+        return response()->json($products);
+    }
+    public function show(Product $product)
+    {
+        return response()->json($product);
     }
 
     //show create product page
-    public function create() {
+    public function create()
+    {
         return view('products.create');
     }
 
     //store a product in db
-    public function store(Request $request) {
+    public function store(Request $request)
+    {
         $rules = [
             'name' => 'required|string|min:5',
             'sku' => 'required|string|min:3',
@@ -36,8 +39,8 @@ class ProductController extends Controller
 
         $validator = Validator::make($request->all(), $rules);
 
-        if($validator->fails()) {
-            return redirect()->route('products.create')->withInput()->withErrors($validator);
+        if ($validator->fails()) {
+            return response()->json(['errors' => $validator->errors()], 422);
         }
 
         //we will store product in db
@@ -46,11 +49,11 @@ class ProductController extends Controller
         $product->sku = $request->sku;
         $product->price = $request->price;
         $product->description = $request->description;
-        if($request->hasFile('image')) {
+        if ($request->hasFile('image')) {
             //here we will store image
-            $imageFile = $request->file('image'); 
-            $ext = $imageFile->getClientOriginalExtension(); 
-            $imageName = time().'.'.$ext; //unique image name
+            $imageFile = $request->file('image');
+            $ext = $imageFile->getClientOriginalExtension();
+            $imageName = time() . '.' . $ext; //unique image name
 
             //save images to products directort 
             $imageFile->move(public_path('uploads/products'), $imageName);
@@ -59,78 +62,50 @@ class ProductController extends Controller
             $product->image = $imageName;
         }
         $product->save();
-
-
-
-        return redirect()->route('products.index')->with('success', 'Product added successfully.');
+        return response()->json($product, 201);
     }
-
-    public function edit($id){
-        $product = Product::findOrFail($id);
-        return view('products.edit', [
-            'product' => $product
-        ]);
-
-    }
-
-    public function update($id, Request $request) {
-        $product = Product::findOrFail($id);
-
-
+    // public function update($id, Request $request)
+    public function update(Request $request, Product $product)
+    {
         $rules = [
             'name' => 'required|string|min:5',
             'sku' => 'required|string|min:3',
             'price' => 'required|numeric|min:0',
-            'image' => 'required'
-
+            'description' => 'nullable',
+            'image' => 'sometimes|required|image|mimes:jpeg,png,jpg,gif,svg|max:2048',
         ];
-
         $validator = Validator::make($request->all(), $rules);
-
-        if($validator->fails()) {
-            return redirect()->route('products.edit', $product->id)->withInput()->withErrors($validator);
+        if ($validator->fails()) {
+            return response()->json(['errors' => $validator->errors()], 422);
         }
-
-        //we will update product in db
         $product->name = $request->name;
         $product->sku = $request->sku;
         $product->price = $request->price;
         $product->description = $request->description;
-        if($request->hasFile('image')) {
-            //delete old image
-            File::delete(public_path('uploads/products/'.$product->image));
-            //here we will update image
-            $imageFile = $request->file('image'); 
-            $ext = $imageFile->getClientOriginalExtension(); 
-            $imageName = time().'.'.$ext; //unique image name
 
-            //save images to products directory
+        if ($request->hasFile('image')) {
+            if ($product->image) {
+                File::delete(public_path('uploads/products/' . $product->image));
+            }
+
+            $imageFile = $request->file('image');
+            $imageName = time() . '.' . $imageFile->getClientOriginalExtension();
             $imageFile->move(public_path('uploads/products'), $imageName);
-
-            //save image name in database
             $product->image = $imageName;
         }
+
         $product->save();
 
-
-
-        return redirect()->route('products.index')->with('success', 'Product updated successfully.');
-
-
+        return response()->json($product);
     }
 
-    public function destroy($id){
-        $product = Product::findOrFail($id);
+    public function destroy(Product $product)
+    {
+        if ($product->image) {
+            File::delete(public_path('uploads/products/' . $product->image));
+        }
 
-        //delete image
-        File::delete(public_path('uploads/products/'.$product->image));
-
-        //delete product 
         $product->delete();
-
-        return redirect()->route('products.index')->with('success', 'Product deleted successfully.');
-
-
+        return response()->json(null, 204);
     }
-
 }
